Class Changes
I used classes to represent object (data entities), this exposed variables to access as they are not protected. I preceded all the variables with _ but this caused problems when accessing the variables in the views. Hence, all the classes were changed back to no underscore. In addition, the constructor had no duplication and was changed to using the setters below.

Factory.js added to provide access to REST. Initial file modified to conform to ES6 format.

There are 4 methods created for accessing data in each entity. More may be needed as I get further into the project.

Signed-off-by: OisinODonnell <oisinodonnell@hotmail.com>