Unit Testing Angular
There were a number of things that needed to come in in order to set up a testing suite.

A 'test' directory was created and a 'unit' directory within that.

bower package manager to bring in angular-mocks

karma was set up as the test runner. Requirejs came in with this. Karma is configured in the karma.conf.js file within 'test' directory.

Karma is the runner that uses Jasmine to test the code.

Npm install was used to bring in all of the development dependencies (viewable in the package.json file). There are now a lot more than are needed in here.

PhantomJS - browser used for testing, in conjunction with Karma, Jasmine and RequireJS

A lot of issues were hit at the start because neither PhantomJS nor Karma were capable of handling ES6/2015. This would have needed for all the source code to change. Instead babel-loader is now used to transpile the code back to ES5. babel-loader is used by webpack.

Webpack is brought into the karma.conf.js file as a preprocessor (along with Babel). Require is used to bring in the webpackConfig file that is in the root of the project. The webpackConfig file is not auto-generated and needed to be brought in, provided an entry-point whereby it could generate all the required code from when running the tests.

Other configurations of the webpackConfig file

For the webpackConfig to work, in the terminal it needed to be passed the webpackConfig file to set itself up initially (webpack --config webpack.config.js), prior to running it through 'karma start app/js/test/karma.conf.js'.

Karma is also set to watch for changes in the code and repeat all the tests during development.

Signed-off-by: OisinODonnell <oisinodonnell@hotmail.com>